// src/types/generated/types.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>
}
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>
}
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  DateTime: any
  Upload: any
}
// ...
export type IUseditem = {
  __typename?: "Useditem"
  _id: Scalars["ID"]
  buyer?: Maybe<IUser>
  contents: Scalars["String"]
  createdAt: Scalars["DateTime"]
  deletedAt?: Maybe<Scalars["DateTime"]>
  images: Array<Scalars["String"]>
  name: Scalars["String"]
  pickedCount?: Maybe<Scalars["Int"]>
  price: Scalars["Int"]
  remarks: Scalars["String"]
  seller?: Maybe<IUser>
  soldAt?: Maybe<Scalars["DateTime"]>
  tags?: Maybe<Array<Scalars["String"]>>
  updatedAt: Scalars["DateTime"]
  useditemAddress?: Maybe<IUseditemAddress>
}

export type IUseditemAddress = {
  __typename?: "UseditemAddress"
  _id: Scalars["ID"]
  address?: Maybe<Scalars["String"]>
  addressDetail?: Maybe<Scalars["String"]>
  createdAt: Scalars["DateTime"]
  deletedAt?: Maybe<Scalars["DateTime"]>
  lat?: Maybe<Scalars["Float"]>
  lng?: Maybe<Scalars["Float"]>
  updatedAt: Scalars["DateTime"]
  zipcode?: Maybe<Scalars["String"]>
}
// ...
export type IUser = {
  __typename?: "User"
  _id: Scalars["ID"]
  createdAt: Scalars["DateTime"]
  deletedAt?: Maybe<Scalars["DateTime"]>
  email: Scalars["String"]
  name: Scalars["String"]
  picture?: Maybe<Scalars["String"]>
  updatedAt: Scalars["DateTime"]
  userPoint?: Maybe<IUserPoint>
}

export type IUserPoint = {
  __typename?: "UserPoint"
  _id: Scalars["ID"]
  amount: Scalars["Int"]
  createdAt: Scalars["DateTime"]
  deletedAt?: Maybe<Scalars["DateTime"]>
  updatedAt: Scalars["DateTime"]
  user: IUser
}